+++ common/src/main/java/net/satisfy/bakery/core/block/CakeStandBlock.java
@@ -8,7 +8,6 @@ import net.minecraft.core.component.DataComponents;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.InteractionHand;
-import net.minecraft.world.InteractionResult;
 import net.minecraft.world.ItemInteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.BlockItem;
@@ -20,6 +19,7 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.BlockHitResult;
+import net.minecraft.world.phys.shapes.BooleanOp;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraft.world.phys.shapes.VoxelShape;
@@ -35,21 +35,17 @@ public class CakeStandBlock extends StorageBlock {
 
     public static VoxelShape makeShape() {
         VoxelShape shape = Shapes.empty();
-        shape = Shapes.or(shape, Shapes.box(0.125, 0.5000625, 0.125, 0.125, 1.0000625, 0.875));
-        shape = Shapes.or(shape, Shapes.box(0.125, 0.5000625, 0.125, 0.875, 1.0000625, 0.125));
-        shape = Shapes.or(shape, Shapes.box(0.125, 0.5000625, 0.875, 0.875, 1.0000625, 0.875));
-        shape = Shapes.or(shape, Shapes.box(0.125, 1.0000625, 0.125, 0.875, 1.0000625, 0.875));
-        shape = Shapes.or(shape, Shapes.box(0.875, 0.5000625, 0.125, 0.875, 1.0000625, 0.875));
-        shape = Shapes.or(shape, Shapes.box(0.4375, 1, 0.4375, 0.5625, 1.0625, 0.5625));
-        shape = Shapes.or(shape, Shapes.box(0.4375, 0.125, 0.4375, 0.5625, 0.4375, 0.5625));
-        shape = Shapes.or(shape, Shapes.box(0.25, 0.375, 0.25, 0.75, 0.4375, 0.75));
-        shape = Shapes.or(shape, Shapes.box(0.125, 0.4375, 0.125, 0.875, 0.5, 0.875));
-        shape = Shapes.or(shape, Shapes.box(0.375, 0.0625, 0.375, 0.625, 0.1875, 0.625));
-        shape = Shapes.or(shape, Shapes.box(0.875, 0.4375, 0.125, 0.9375, 0.5, 0.875));
-        shape = Shapes.or(shape, Shapes.box(0.0625, 0.4375, 0.125, 0.125, 0.5, 0.875));
-        shape = Shapes.or(shape, Shapes.box(0.125, 0.4375, 0.0625, 0.875, 0.5, 0.125));
-        shape = Shapes.or(shape, Shapes.box(0.125, 0.4375, 0.875, 0.875, 0.5, 0.9375));
-        shape = Shapes.or(shape, Shapes.box(0.3125, 0, 0.3125, 0.6875, 0.0625, 0.6875));
+        shape = Shapes.join(shape, Shapes.box(0.125, 0.5000625, 0.125, 0.125, 1.0000625, 0.875), BooleanOp.OR);
+        shape = Shapes.join(shape, Shapes.box(0.125, 0.5000625, 0.125, 0.875, 1.0000625, 0.125), BooleanOp.OR);
+        shape = Shapes.join(shape, Shapes.box(0.125, 0.5000625, 0.875, 0.875, 1.0000625, 0.875), BooleanOp.OR);
+        shape = Shapes.join(shape, Shapes.box(0.125, 1.0000625, 0.125, 0.875, 1.0000625, 0.875), BooleanOp.OR);
+        shape = Shapes.join(shape, Shapes.box(0.875, 0.5000625, 0.125, 0.875, 1.0000625, 0.875), BooleanOp.OR);
+        shape = Shapes.join(shape, Shapes.box(0.4375, 1, 0.4375, 0.5625, 1.0625, 0.5625), BooleanOp.OR);
+        shape = Shapes.join(shape, Shapes.box(0.4375, 0.125, 0.4375, 0.5625, 0.4375, 0.5625), BooleanOp.OR);
+        shape = Shapes.join(shape, Shapes.box(0.25, 0.375, 0.25, 0.75, 0.4375, 0.75), BooleanOp.OR);
+        shape = Shapes.join(shape, Shapes.box(0.125, 0.4375, 0.125, 0.875, 0.5, 0.875), BooleanOp.OR);
+        shape = Shapes.join(shape, Shapes.box(0.375, 0.0625, 0.375, 0.625, 0.1875, 0.625), BooleanOp.OR);
+        shape = Shapes.join(shape, Shapes.box(0.3125, 0, 0.3125, 0.6875, 0.0625, 0.6875), BooleanOp.OR);
         return shape;
     }