+++ common/src/main/java/net/satisfy/bakery/core/block/entity/StorageBlockEntity.java
@@ -15,7 +15,6 @@ import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.satisfy.bakery.core.registry.EntityTypeRegistry;
 import net.satisfy.farm_and_charm.core.util.GeneralUtil;
-import org.jetbrains.annotations.NotNull;
 
 public class StorageBlockEntity extends BlockEntity {
     private int size;