+++ common/src/main/java/net/satisfy/bakery/core/block/entity/StreetSignBlockEntity.java
@@ -5,12 +5,11 @@ import net.minecraft.core.HolderLookup;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.protocol.Packet;
-import net.minecraft.network.protocol.game.ClientboundBlockEntityDataPacket;
 import net.minecraft.network.protocol.game.ClientGamePacketListener;
+import net.minecraft.network.protocol.game.ClientboundBlockEntityDataPacket;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.satisfy.bakery.core.registry.EntityTypeRegistry;
-import org.jetbrains.annotations.NotNull;
 
 public class StreetSignBlockEntity extends BlockEntity {
     private final Component[] text = new Component[]{Component.literal(""), Component.literal(""), Component.literal("")};