+++ common/src/main/java/net/satisfy/bakery/core/network/SetStreetSignTextPacket.java
@@ -1,7 +1,6 @@
 package net.satisfy.bakery.core.network;
 
 import net.minecraft.core.BlockPos;
-import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.codec.StreamCodec;